Feat: Comite solicitudes list ready, fixes on solicitudes, analisis grupales started

+++ lib/src/datasource/solicitudes/hn/solicitudes/asalariado/solicitud_asalariado_hn.dart
@@ -34,7 +34,7 @@ class SolicitudAsalariadoHn {
   final String monedaCodigo;
   final String propositoCodigo;
   final String frecuenciaCodigo;
-  final int cuota;
+  final double cuota;
   final String sectorCodigo;
   final int personasACargo;
   final String estadoCivilCodigo;
